+++ src/items/items.pri
@@ -1,0 +1,10 @@
+SRC_GROUPS += ITEMS
+ITEMS_BASE = src/items
+
+ITEMS_HEADERS += gripitem.h   edgeitem.h   markeritem.h
+ITEMS_SOURCES += gripitem.cpp edgeitem.cpp markeritem.cpp
+
+ITEMS_HEADERS += meshitem.h   meshpolygon.h   polylineitem.h
+ITEMS_SOURCES += meshitem.cpp meshpolygon.cpp polylineitem.cpp
+
+ITEMS_HEADERS += meshrenderdata.h
