+++ src/render/penstrokerenderer.cpp
