+++ src/render/penstrokerenderer.h
