+++ src/render/render.pri
@@ -1,0 +1,8 @@
+SRC_GROUPS += RENDER
+RENDER_BASE = src/render
+
+RENDER_HEADERS += abstractmeshrenderer.h   propertypanel.h
+RENDER_SOURCES += abstractmeshrenderer.cpp propertypanel.cpp
+
+RENDER_HEADERS += meshgradientrenderer.h   penstrokerenderer.h
+RENDER_SOURCES += meshgradientrenderer.cpp penstrokerenderer.cpp
