+++ src/tools/tools.pri
@@ -1,0 +1,8 @@
+SRC_GROUPS += TOOLS
+TOOLS_BASE = src/tools
+
+TOOLS_HEADERS += tool.h
+TOOLS_SOURCES += tool.cpp
+
+TOOLS_HEADERS += movevertex.h   moveedge.h   color.h   split.h
+TOOLS_SOURCES += movevertex.cpp moveedge.cpp color.cpp split.cpp
